fix API
